+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.x [TBD]
 - Added `tools/run_cases.py` with tools to run different `tech_config` cases from a spreadsheet, with new docs page to describe: docs/user_guide/how_to_run_several_cases_in_sequence.md
+- Added capability for user-defined finance models in the H2Integrate framework
 
 ## 0.4.0 [October 1, 2025]